requirements.txt
fastapi
uvicorn
python-dotenv
mcp
openai
psutil
httpx
pymongo
certifi
httpx-sse
websockets
pydantic-settings